fix(monetary): corregir tests y servicios después de migración a integer
CORRECCIONES APLICADAS:

1. Tests actualizados:
   - CashierServicesTest: cambiar ->toBe(20000.0) a ->toBe(20000)
   - SyncAdapterTest: cambiar ->toBe(11900.0) a ->toBe(11900)
   - Los tests ahora esperan int en lugar de float

2. TipPayoutService:
   - received, paid, pending: (float) → (int)
   - Estos SON montos monetarios (propinas recibidas/pagadas)
   - Error de clasificación corregido

3. KitchenQueueService:
   - total_amount: (float) → (int)
   - Este ES un monto monetario (total de órdenes en cocina)
   - Error de clasificación corregido

RESULTADO:
✅ Todos los tests de Cashier pasan
✅ Todos los tests de Sync pasan
✅ Modelo monetario 100% integer consistente

+++ app/Modules/Cashier/Domain/Services/TipPayoutService.php
@@ -33,9 +33,9 @@ class TipPayoutService
         $paid = $this->getWaiterPaid($sessionId, $waiterId);
 
         return [
-            'received' => (float) $received,  // Float justificado: no es monto monetario
-            'paid' => (float) $paid,  // Float justificado: no es monto monetario
-            'pending' => max(0, (float) $received - (float) $paid),  // Float justificado: no es monto monetario
+            'received' => (int) $received,
+            'paid' => (int) $paid,
+            'pending' => max(0, (int) $received - (int) $paid),
         ];
     }
